tests/wpt/web-platform-tests/css/css-text/parsing/text-indent-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Module Test: parsing text-indent with valid values</title>
<link rel="help" href="https://www.w3.org/TR/css-text-3/#propdef-text-indent">
<meta name="assert" content="text-indent supports the full grammar '[ <length-percentage> ] && hanging? && each-line?'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("text-indent", "10px");
test_valid_value("text-indent", "20%");
test_valid_value("text-indent", "calc(2em + 3ex)");
test_valid_value("text-indent", "calc(50% + 60px)");
test_valid_value("text-indent", "-30px");
test_valid_value("text-indent", "-40%");

test_valid_value("text-indent", "10px hanging");
test_valid_value("text-indent", "20% each-line");
test_valid_value("text-indent", "calc(2em + 3ex) hanging each-line");

test_valid_value("text-indent", "hanging calc(50% + 60px)", "calc(50% + 60px) hanging");
test_valid_value("text-indent", "each-line 10px", "10px each-line");
test_valid_value("text-indent", "hanging 20% each-line", "20% hanging each-line");
test_valid_value("text-indent", "each-line hanging calc(2em + 3ex)", "calc(2em + 3ex) hanging each-line");
test_valid_value("text-indent", "calc(50% + 60px) each-line hanging", "calc(50% + 60px) hanging each-line");
</script>
</body>
</html>
